requirements.txt
APScheduler==3.6.3
asgiref==3.2.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.1
django-connection-url==0.1.2
django-cors-headers==3.5.0
djangorestframework==3.11.1
idna==2.10
psycopg2-binary==2.8.6
pyaspeller==0.2.0
pycparser==2.20
pyTelegramBotAPI==3.7.6
python-telegram-bot==13.2
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
